The database initialization logic was refactored to allow for explicit initialization and testing configurations, including moving the store synchronization logic into the initialization function, and test dependencies now install both requirements files before tests.
